[JENKINS-45333] Code refactor of hook to generate effective pom for BO and DP
